requirements.txt
pytest~=8.3.4
allure-pytest==2.13.5
selene~=2.0.0rc9
allure-python-commons~=2.13.5
selenium~=4.27.1